// Server/src/app/modules/booking/booking.interface.ts
//* User -> Booking(pending) -> Payment(unpaid) -> SSLCommerz -> Booking Update = Confirm -> Payment Update -> Paid

import { Types } from "mongoose";

export enum BOOKING_STATUS {
  PENDING = "PENDING",
  CANCEL = "CANCEL",
  COMPLETE = "COMPLETE",
  "FAILED" = "FAILED",
}

export interface IBooking {
  user: Types.ObjectId;
  tour: Types.ObjectId;
  payment?: Types.ObjectId;
  guestCount: number;
  status: BOOKING_STATUS;
}